+++ config/task/command_test.go
@@ -2,10 +2,12 @@ package task
 
 import (
 	"os"
+	"os/exec"
+	"path/filepath"
+	"strings"
 	"testing"
 
 	"github.com/google/go-cmp/cmp"
-	"github.com/pkg/errors"
 	yaml "gopkg.in/yaml.v2"
 )
 
@@ -39,6 +41,15 @@ func TestCommand_UnmarshalYAML(t *testing.T) {
 				Print: "echo example",
 			},
 		},
+		{
+			"many-fields",
+			`{do: dovalue, print: printvalue, dir: dirvalue}`,
+			Command{
+				Do:    "dovalue",
+				Print: "printvalue",
+				Dir:   "dirvalue",
+			},
+		},
 	}
 
 	for _, tt := range tests {
@@ -53,6 +64,77 @@ func TestCommand_UnmarshalYAML(t *testing.T) {
 				t.Errorf("mismatched values:\n%s", diff)
 			}
 		})
+	}
+}
+
+// TestCommand_exec_helper is a helper test that is called when mocking exec.
+//
+// The following environment variables can configure this function:
+//
+// - TUSK_WANT_TEST_COMMAND: Set to "1" to run this function.
+// - TUSK_TEST_COMMAND_ARGS: Set to a comma-separated list of expected command
+//   arguments.
+// - TUSK_TEST_COMMAND_DIR: Set to the expected directory
+func TestCommand_exec_helper(t *testing.T) {
+	if os.Getenv("TUSK_WANT_TEST_COMMAND") != "1" {
+		return
+	}
+
+	wantArgs := strings.Split(os.Getenv("TUSK_TEST_COMMAND_ARGS"), ",")
+	args := os.Args
+	for len(args) > 0 {
+		if args[0] == "--" {
+			args = args[1:]
+			break
+		}
+		args = args[1:]
+	}
+
+	if diff := cmp.Diff(wantArgs, args); diff != "" {
+		t.Errorf("arguments differ:\n%s", diff)
+	}
+
+	dir, err := os.Getwd()
+	if err != nil {
+		t.Fatal("failed to get working dir: ", err)
+	}
+
+	wantDir := os.Getenv("TUSK_TEST_COMMAND_DIR")
+	if wantDir != "" && dir != wantDir {
+		t.Errorf("want working dir %s, got %s", wantDir, dir)
+	}
+}
+
+func TestCommand_exec(t *testing.T) {
+	wantCommand := "echo hello world"
+	wantArgs := strings.Join([]string{getShell(), "-c", wantCommand}, ",")
+
+	wd, err := os.Getwd()
+	if err != nil {
+		t.Fatal(err)
+	}
+	wantDir := filepath.Dir(wd)
+
+	command := Command{
+		Do:  wantCommand,
+		Dir: "..",
+	}
+
+	execCommand = func(name string, arg ...string) *exec.Cmd {
+		cs := []string{"-test.run=TestCommand_exec_helper", "--", name}
+		cs = append(cs, arg...)
+		cmd := exec.Command(os.Args[0], cs...) // nolint: gosec
+		cmd.Env = []string{
+			"TUSK_WANT_TEST_COMMAND=1",
+			"TUSK_TEST_COMMAND_ARGS=" + wantArgs,
+			"TUSK_TEST_COMMAND_DIR=" + wantDir,
+		}
+		return cmd
+	}
+	defer func() { execCommand = exec.Command }()
+
+	if err := command.exec(); err != nil {
+		t.Fatal(err)
 	}
 }
 
@@ -111,24 +193,6 @@ func TestCommandList_UnmarshalYAML(t *testing.T) {
 				t.Errorf("mismatched values:\n%s", diff)
 			}
 		})
-	}
-}
-
-func TestExecCommand(t *testing.T) {
-	command := "exit 0"
-
-	if err := execCommand(command); err != nil {
-		t.Fatalf(`execCommand("%s"): unexpected err: %s`, command, err)
-	}
-}
-
-func TestExecCommand_error(t *testing.T) {
-	command := "exit 1"
-	errExpected := errors.New("exit status 1")
-	if err := execCommand(command); err.Error() != errExpected.Error() {
-		t.Fatalf(`execCommand("%s"): expected error "%s", actual "%s"`,
-			command, errExpected, err,
-		)
 	}
 }